Fix upgrade
